# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk Sahabat Kurban",
        "title": "Sahabat Kurban",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Restu Adi Wahyujati",
            "email": "restu@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/patungan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua data patungan (admin/panitia), atau hanya milik user sendiri",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Get all patungan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User menambahkan dirinya ke dalam patungan hewan kurban",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Create new patungan hewan",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePekurbanHewanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/patungan/hewan/{hewan_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua pekurban berdasarkan ID hewan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Get patungan by hewan_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hewan ID",
                        "name": "hewan_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/patungan/pekurban/{pekurban_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ambil semua data patungan berdasarkan ID pekurban",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Get patungan by pekurban_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pekurban ID",
                        "name": "pekurban_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/patungan/{pekurban_id}/{hewan_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update porsi kontribusi user dalam patungan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Update patungan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pekurban ID",
                        "name": "pekurban_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hewan ID",
                        "name": "hewan_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePekurbanHewanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hapus relasi patungan pekurban dengan hewan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patungan"
                ],
                "summary": "Delete patungan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pekurban ID",
                        "name": "pekurban_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hewan ID",
                        "name": "hewan_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePekurbanHewanRequest": {
            "type": "object",
            "required": [
                "hewan_id",
                "jumlah_orang",
                "pekurban_id"
            ],
            "properties": {
                "hewan_id": {
                    "type": "string"
                },
                "jumlah_orang": {
                    "type": "integer",
                    "maximum": 7
                },
                "pekurban_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePekurbanHewanRequest": {
            "type": "object",
            "required": [
                "jumlah_orang"
            ],
            "properties": {
                "jumlah_orang": {
                    "type": "integer",
                    "maximum": 7
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" + a space + your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}